requirements.txt
# VerifyPulse Backend Dependencies

# Core Framework
fastapi==0.121.3
uvicorn[standard]==0.38.0

# CLI
typer==0.20.0
rich==14.2.0

# Data & Validation
pydantic==2.12.4
python-dotenv==1.0.1

# HTTP Client
requests==2.32.5
httpx==0.27.0

# Redis
redis==6.4.0
redisvl==0.11.1

# Skyflow
skyflow==2.0.0
